Move: parser 위치 변경, 불필요한 import 제거
