+++ src/main/java/christmas/service/EventFinder.java
@@ -1,7 +1,6 @@
 package christmas.service;
 
 import christmas.domain.event.EventDetail;
-import christmas.domain.event.MatchingEvents;
 import christmas.domain.orders.Orders;
 import christmas.domain.visitingDate.VisitingDate;
