+++ src/main/java/christmas/view/InputView.java
@@ -2,6 +2,7 @@ package christmas.view;
 
 import camp.nextstep.edu.missionutils.Console;
 import christmas.dto.OrderItemDto;
+import christmas.validationUtils.Parser;
 
 import java.util.List;
